Refactor into 3 tool states
