--- S2VX.Game/Editor/ToolState/CameraToolDragState.cs
@@ -1,7 +1,0 @@
-﻿namespace S2VX.Game.Editor.ToolState {
-    public enum CameraToolDragState {
-        Move,
-        Scale,
-        Rotate
-    }
-}
